Translate/requirements.txt
pyodbc
requests
uuid
python-dotenv